# perf: Parallelize node filtering