--- a/pkg/controllers/provisioning/scheduling/nodeclaim.go
+++ b/pkg/controllers/provisioning/scheduling/nodeclaim.go
@@ -109,32 +109,35 @@ func NewNodeClaim(
 	}
 }
 
-func (n *NodeClaim) Add(ctx context.Context, pod *corev1.Pod, podData *PodData) error {
+// CanAdd returns whether the pod can be added to the NodeClaim
+// based on the taints/tolerations, host port compatibility,
+// requirements, resources, reserved capacity reservations, and topology requirements
+func (n *NodeClaim) CanAdd(ctx context.Context, pod *corev1.Pod, podData *PodData) (updateRequirements scheduling.Requirements, updatedInstanceTypes []*cloudprovider.InstanceType, offeringsToReserve []*cloudprovider.Offering, err error) {
 	// Check Taints
 	if err := scheduling.Taints(n.Spec.Taints).ToleratesPod(pod); err != nil {
-		return err
+		return nil, nil, nil, err
 	}
 
 	// exposed host ports on the node
 	hostPorts := scheduling.GetHostPorts(pod)
 	if err := n.hostPortUsage.Conflicts(pod, hostPorts); err != nil {
-		return fmt.Errorf("checking host port usage, %w", err)
+		return nil, nil, nil, fmt.Errorf("checking host port usage, %w", err)
 	}
 	nodeClaimRequirements := scheduling.NewRequirements(n.Requirements.Values()...)
 
 	// Check NodeClaim Affinity Requirements
 	if err := nodeClaimRequirements.Compatible(podData.Requirements, scheduling.AllowUndefinedWellKnownLabels); err != nil {
-		return fmt.Errorf("incompatible requirements, %w", err)
+		return nil, nil, nil, fmt.Errorf("incompatible requirements, %w", err)
 	}
 	nodeClaimRequirements.Add(podData.Requirements.Values()...)
 
 	// Check Topology Requirements
 	topologyRequirements, err := n.topology.AddRequirements(pod, n.NodeClaimTemplate.Spec.Taints, podData.StrictRequirements, nodeClaimRequirements, scheduling.AllowUndefinedWellKnownLabels)
 	if err != nil {
-		return err
+		return nil, nil, nil, err
 	}
 	if err = nodeClaimRequirements.Compatible(topologyRequirements, scheduling.AllowUndefinedWellKnownLabels); err != nil {
-		return err
+		return nil, nil, nil, err
 	}
 	nodeClaimRequirements.Add(topologyRequirements.Values()...)
 
@@ -145,28 +148,33 @@ func (n *NodeClaim) Add(ctx context.Context, pod *corev1.Pod, podData *PodData) 
 	if err != nil {
 		// We avoid wrapping this err because calling String() on InstanceTypeFilterError is an expensive operation
 		// due to calls to resources.Merge and stringifying the nodeClaimRequirements
-		return err
+		return nil, nil, nil, err
 	}
-
-	reservedOfferings, err := n.reserveOfferings(ctx, remaining, nodeClaimRequirements)
+	ofs, err := n.offeringsToReserve(ctx, remaining, nodeClaimRequirements)
 	if err != nil {
-		return err
+		return nil, nil, nil, err
 	}
+	return nodeClaimRequirements, remaining, ofs, nil
+}
 
+// Add updates the NodeClaim to schedule the pod to this NodeClaim, updating
+// the NodeClaim with new requirements, instance types, and offerings to reserve
+// based on the pod scheduling
+func (n *NodeClaim) Add(pod *corev1.Pod, podData *PodData, nodeClaimRequirements scheduling.Requirements, instanceTypes []*cloudprovider.InstanceType, offeringsToReserve []*cloudprovider.Offering) {
 	// Update node
 	n.Pods = append(n.Pods, pod)
-	n.InstanceTypeOptions = remaining
-	n.Spec.Resources.Requests = requests
+	n.InstanceTypeOptions = instanceTypes
+	n.Spec.Resources.Requests = resources.Merge(n.Spec.Resources.Requests, podData.Requests)
 	n.Requirements = nodeClaimRequirements
 	n.topology.Record(pod, n.NodeClaim.Spec.Taints, nodeClaimRequirements, scheduling.AllowUndefinedWellKnownLabels)
-	n.hostPortUsage.Add(pod, hostPorts)
-	n.releaseReservedOfferings(n.reservedOfferings, reservedOfferings)
-	n.reservedOfferings = reservedOfferings
-	return nil
+	n.hostPortUsage.Add(pod, scheduling.GetHostPorts(pod))
+	n.reservationManager.Reserve(n.hostname, offeringsToReserve...)
+	n.releaseReservedOfferings(n.reservedOfferings, offeringsToReserve)
+	n.reservedOfferings = offeringsToReserve
 }
 
 // releaseReservedOfferings releases all offerings which are present in the current reserved offerings, but are not
-// present in the updated reserved offerings.
+// present in the updated reserved offerings
 func (n *NodeClaim) releaseReservedOfferings(current, updated cloudprovider.Offerings) {
 	updatedIDs := sets.New[string]()
 	for _, o := range updated {
@@ -184,7 +192,7 @@ func (n *NodeClaim) releaseReservedOfferings(current, updated cloudprovider.Offe
 // to reserve compatible offerings when some were available.
 //
 //nolint:gocyclo
-func (n *NodeClaim) reserveOfferings(
+func (n *NodeClaim) offeringsToReserve(
 	ctx context.Context,
 	instanceTypes []*cloudprovider.InstanceType,
 	nodeClaimRequirements scheduling.Requirements,
@@ -209,7 +217,7 @@ func (n *NodeClaim) reserveOfferings(
 			// Note that reservation is an idempotent operation - if we have previously successfully reserved an offering for
 			// this host, this operation is guaranteed to succeed. We may also succeed to make reservations for offerings which
 			// failed in previous iterations if other NodeClaims have released them since the last attempt.
-			if n.reservationManager.Reserve(n.hostname, o) {
+			if n.reservationManager.CanReserve(n.hostname, o) {
 				reservedOfferings = append(reservedOfferings, o)
 			}
 		}