--- a/pkg/controllers/provisioning/scheduling/reservationmanager.go
+++ b/pkg/controllers/provisioning/scheduling/reservationmanager.go
@@ -72,6 +72,9 @@ func (rm *ReservationManager) CanReserve(hostname string, offering *cloudprovide
 func (rm *ReservationManager) Reserve(hostname string, offerings ...*cloudprovider.Offering) {
 	for _, of := range offerings {
 		rm.capacity[of.ReservationID()] -= 1
+		if rm.capacity[of.ReservationID()] < 0 {
+			panic(fmt.Sprintf("attempted to over-reserve an offering with reservation id %q", of.ReservationID()))
+		}
 		if _, ok := rm.reservations[hostname]; !ok {
 			rm.reservations[hostname] = sets.New[string]()
 		}