… kwok/charts/templates/deployment.yaml
… pkg/controllers/state/suite_test.go
… pkg/operator/options/options.go
… pkg/test/options.go